Adds first pass Main component
